+++ tests/test_archive_api.py
@@ -84,33 +84,6 @@ def test_compress_archive(tmp_archive):
     assert not os.path.isfile(tar_filename)
 
 
-def test_extract_file(tmp_archive):
-    from dtool.archive import extract_file
-
-    base_dir, tar_gz_filename = os.path.split(tmp_archive)
-    file_prefix, ext = tar_gz_filename.split(".", 1)
-
-    expected_path = os.path.join(base_dir,
-                                 file_prefix,
-                                 "README.yml")
-    expected_path = os.path.abspath(expected_path)
-    readme_path = extract_file(tmp_archive, "README.yml")
-    assert readme_path == expected_path
-    assert os.path.isfile(readme_path)
-
-    # Remove the extracted file and unzip the tarball.
-    os.unlink(readme_path)
-    unzip_command = ["gunzip", tmp_archive]
-    subprocess.call(unzip_command)
-    tarball_path, _ = tmp_archive.rsplit(".", 1)
-    assert os.path.isfile(tarball_path)
-
-    # Test that the extract_file method works on unzipped tarballs too.
-    readme_path = extract_file(tarball_path, "README.yml")
-    assert readme_path == expected_path
-    assert os.path.isfile(readme_path)
-
-
 def test_archive_verify_all(tmp_archive):
     from dtool.archive import verify_all